New shortcode providers, update tests
