+++ tests/php/ImageTest.php
@@ -266,60 +266,5 @@ abstract class ImageTest extends SapphireTest
         $this->assertEquals($resampled2->TestProperty, $testString);
     }
 
-    public function testShortcodeHandlerFallsBackToFileProperties()
-    {
-        $image = $this->objFromFixture(Image::class, 'imageWithTitle');
-        $parser = new ShortcodeParser();
-        $parser->register('image', array(Image::class, 'handle_shortcode'));
 
-        $this->assertEquals(
-            sprintf(
-                '<img src="%s" alt="%s">',
-                $image->Link(),
-                $image->Title
-            ),
-            $parser->parse(sprintf('[image id=%d]', $image->ID))
-        );
-    }
-
-    public function testShortcodeHandlerUsesShortcodeProperties()
-    {
-        $image = $this->objFromFixture(Image::class, 'imageWithTitle');
-        $parser = new ShortcodeParser();
-        $parser->register('image', array(Image::class, 'handle_shortcode'));
-
-        $this->assertEquals(
-            sprintf(
-                '<img src="%s" alt="Alt content" title="Title content">',
-                $image->Link()
-            ),
-            $parser->parse(
-                sprintf(
-                    '[image id="%d" alt="Alt content" title="Title content"]',
-                    $image->ID
-                )
-            )
-        );
-    }
-
-    public function testShortcodeHandlerAddsDefaultAttributes()
-    {
-        $image = $this->objFromFixture(Image::class, 'imageWithoutTitle');
-        $parser = new ShortcodeParser();
-        $parser->register('image', array(Image::class, 'handle_shortcode'));
-
-        $this->assertEquals(
-            sprintf(
-                '<img src="%s" alt="%s">',
-                $image->Link(),
-                $image->Title
-            ),
-            $parser->parse(
-                sprintf(
-                    '[image id="%d"]',
-                    $image->ID
-                )
-            )
-        );
-    }
 }
